requirements_mac.txt
absl-py==0.9.0
appdirs==1.4.4
asgiref==3.2.10
astor==0.8.1
asyncio==3.4.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==3.2
distlib==0.3.1
Django==3.0.8
filelock==3.0.12
gast==0.2.2
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
image==1.5.32
imgur-uploader==0.3.0
imgurpython==1.1.7
importlib-metadata==1.7.0
imutils==0.5.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.19.0
oauthlib==3.1.0
opencv-python==3.4.6.27
opt-einsum==3.2.1
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyimgur==0.6.0
pyperclip==1.8.0
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
tensorboard==2.1.1
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-2.1.0-cp37-cp37m-macosx_10_9_x86_64.whl
termcolor==1.1.0
tflite-runtime @ https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-macosx_10_14_x86_64.whl
urllib3==1.25.9
virtualenv==20.0.28
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0